in class BookDaoImpl add implementation for methods:
findBookById and findAllBook. Corrected small mistakes in classes Book, BookDaoImpl

+++ src/main/java/com/zinko/dao/BookDao.java
@@ -8,7 +8,7 @@ public interface BookDao {
 
     void creatBook(Book book);
 
-    Book findBookById(int id);
+    Book findBookById(Long id);
 
     List<Book> findAllBook();
